requirements.txt
Flask
slack-bolt
slack-sdk
python-dotenv
openai
langchain
langchain-openai
langchain-community
aiohttp
hmac
hashlib
